refactor(agence_bio): use DataProcessor
Remove the filter during the API calls on the départements. Agence BIO
API now accepts calls without it. Several hundreds of records has no
départements.
