+++ .gitignore
@@ -4,7 +4,8 @@ __pycache__/
 .ruff_cache/
 
 .env
-**/data_pipelines/proto/*
+local-tmp/
+
 pyproject.toml
 poetry.lock
 .venv
