+++ workflows/data_pipelines/etl/data_fetch_clean/agence_bio.py
@@ -1,5 +1,7 @@
 import pandas as pd
-from dag_datalake_sirene.config import URL_MINIO_AGENCE_BIO
+from dag_datalake_sirene.workflows.data_pipelines.agence_bio.agence_bio_config import (
+    AGENCE_BIO_CONFIG,
+)
 
 
 def get_statut(statuts):
@@ -11,7 +13,7 @@ def get_statut(statuts):
 
 
 def preprocess_agence_bio_data(data_dir):
-    df_agence_bio = pd.read_csv(URL_MINIO_AGENCE_BIO, dtype=str)
+    df_agence_bio = pd.read_csv(AGENCE_BIO_CONFIG.url_minio, dtype=str)
 
     df_agence_bio = df_agence_bio[["siret", "id_bio", "etat_certification"]]
     df_agence_bio = df_agence_bio[df_agence_bio["siret"].str.len() == 14]
